refactor: rename provider to context in components
Replaced `provider` with `context` across multiple components for consistency with established naming conventions. Updated all related imports, exports, file paths, and references accordingly.
