+++ apps/nextjs/src/components/features/inbox/inbox-list-item/inbox-list-item.tsx
@@ -2,7 +2,7 @@ import { memo } from 'react';
 import { useInboxListItem } from '@/components/features/inbox/hooks/use-inbox-list-item';
 import type { FlexProps } from '@/components/ui/flex';
 import { ActivityTypeCode } from '@/store/entities/activity-type';
-import { Context } from './provider';
+import { Context } from './context';
 import { TaskActivity } from './task-activity';
 import { WorkspaceActivity } from './workspace-activity';
