+++ apps/nextjs/src/components/features/inbox/inbox-list-item/task-activity/action-buttons.tsx
@@ -2,7 +2,7 @@ import { memo } from 'react';
 import { useInboxContext } from '@/components/features/inbox/inbox/context';
 import type { FlexProps } from '@/components/ui/flex';
 import { Actions, ArchiveButton, MoveToInboxButton } from '../actions';
-import { useInboxListItemContext } from '../provider';
+import { useInboxListItemContext } from '../context';
 
 type Props = FlexProps;
