+++ apps/nextjs/src/components/features/task-detail/task-detail-body/form/description/description.tsx
@@ -8,8 +8,8 @@ import {
 import { useTask } from '@/store/entities/task';
 import { Content, Label, Row } from '../row';
 import { Container } from './container';
+import { Context } from './context';
 import { Placeholder } from './placeholder';
-import { Context } from './provider';
 import { ToolBar } from './tool-bar';
 
 type Props = {
