+++ apps/nextjs/src/components/features/task-detail/task-detail-body/form/feed-list/feed-list-item/provider/index.ts
@@ -1,1 +1,1 @@
-export * from './provider';
+export * from './context';
