+++ apps/nextjs/src/components/features/task-detail/task-detail-body/form/subtasks/tasks-name/assignee/use-assignee.ts
@@ -1,5 +1,5 @@
 import { useCallback, useMemo, useState } from 'react';
-import { useSubtasksNameContext } from '../provider';
+import { useSubtasksNameContext } from '../context';
 
 export type UseAssignee = {
   onAssigneeOpened: () => void;
