+++ apps/nextjs/src/components/features/task-detail/task-detail-body/form/subtasks/tasks-name/tasks-name.tsx
@@ -6,8 +6,8 @@ import { Stack } from '@/components/ui/stack';
 import { TaskDoneTransition } from '@/components/ui/transitions';
 import { useTask, useTaskCommand } from '@/store/entities/task';
 import { Assignee } from './assignee';
+import { Context, useSubtasksNameContext } from './context';
 import { DueDate } from './due-date';
-import { Context, useSubtasksNameContext } from './provider';
 import { RightItem } from './right-item';
 import { TasksNameCell } from './tasks-name-cell';
 import { TasksNameField } from './tasks-name-field';
