+++ apps/nextjs/src/components/features/task-detail/task-detail-footer/collaborators/collaborators.tsx
@@ -1,9 +1,9 @@
 import { memo } from 'react';
 import { Flex } from '@/components/ui/flex';
 import { Text } from '@/components/ui/text';
+import { Context } from './context';
 import { Input } from './input';
 import { LeaveTask } from './leave-task';
-import { Context } from './provider';
 import { Teammates } from './teammates';
 
 export const Collaborators = memo(function Collaborators(props) {
