+++ apps/nextjs/src/components/features/task-detail/task-detail-footer/comment/input/input.tsx
@@ -4,8 +4,8 @@ import { Flex } from '@/components/ui/flex';
 import { getDefaultDescription } from '@/shared/prosemirror/get-default-description';
 import { Attachments } from './attachments';
 import { Container } from './container';
+import { Context, useInputContext } from './context';
 import { Placeholder } from './placeholder';
-import { Context, useInputContext } from './provider';
 import { ToolBar } from './tool-bar';
 
 const initialValue = JSON.stringify(getDefaultDescription());
