+++ apps/nextjs/src/components/features/task-detail/task-detail-footer/comment/input/tool-bar/tool-bar.tsx
@@ -4,7 +4,7 @@ import { AtMention, Emoji, Format } from '@/components/ui/editor/tool-bar';
 import { Flex } from '@/components/ui/flex';
 import { Stack } from '@/components/ui/stack';
 import { transitions } from '@/styles/transitions';
-import { useInputContext } from '../provider';
+import { useInputContext } from '../context';
 import { Attachment } from './attachment';
 
 export const ToolBar = memo(function ToolBar() {
