+++ apps/nextjs/src/components/features/thumbnail-attachment/image/image.tsx
@@ -1,8 +1,8 @@
 import { Container } from '@/components/features/thumbnail-attachment/container';
+import { useThumbnailAttachmentContext } from '@/components/features/thumbnail-attachment/context';
 import { Menu } from '@/components/features/thumbnail-attachment/menu';
 import { MenuButton } from '@/components/features/thumbnail-attachment/menu-button';
 import { Overlay } from '@/components/features/thumbnail-attachment/overlay';
-import { useThumbnailAttachmentContext } from '@/components/features/thumbnail-attachment/provider';
 import { Tooltip } from '@/components/features/thumbnail-attachment/tooltip';
 import type { FlexProps } from '@/components/ui/flex';
 import { Image as AtomsImage } from '@/components/ui/image';
