+++ apps/nextjs/src/components/features/thumbnail-attachment/tooltip.tsx
@@ -6,7 +6,7 @@ import {
 } from '@/components/ui/tooltip';
 import { useDisclosure } from '@/shared/chakra';
 import { useTaskFile } from '@/store/entities/task-file';
-import { useThumbnailAttachmentContext } from './provider';
+import { useThumbnailAttachmentContext } from './context';
 
 type Props = Omit<TooltipProps, 'content' | 'size'> & {
   taskFileId: string;
